// src/types/document-viewer.types.ts
export type SelectionAction =
  | "copy"
  | "highlight"
  | "underline"
  | "share"
  | "search"
  | "explain";